+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Also, this project likes fast realizes with even one change to get it live short
 ## Next release
 New features and bug fixes will be here.
 - Fix `updateNode()`: don't increment level if no level column.
+- Allow `level` to be `null` if it's virtual.
+- Add more tests.
 
 ## 1.6.1 - 2021-07-07
 - Fix `getAncestors()` for `depth` greater than 1.